refactor classes and other changes
